+++ ReactNativeStudy/Lisa/Week4/map/src/models/navigator.ts
